+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/api/genes/Gene.kt
@@ -22,7 +22,6 @@ import net.minecraft.network.chat.MutableComponent
 import net.minecraft.network.codec.ByteBufCodecs
 import net.minecraft.network.codec.StreamCodec
 import net.minecraft.resources.RegistryFileCodec
-import net.minecraft.resources.RegistryFixedCodec
 import net.minecraft.world.effect.MobEffect
 import net.minecraft.world.effect.MobEffectInstance
 import net.minecraft.world.entity.Entity
@@ -38,9 +37,9 @@ import kotlin.jvm.optionals.getOrNull
 data class Gene(
     val isNegative: Boolean = false,
     val isHidden: Boolean = false,
+    val isMutation: Boolean = false,
     val dnaPointsRequired: Int = 1,
     val allowedEntities: HolderSet<EntityType<*>> = AnyHolderSet(BuiltInRegistries.ENTITY_TYPE.asLookup()),
-    val mutatesInto: Optional<Holder<Gene>> = Optional.empty(),
     val potionDetails: Optional<PotionDetails> = Optional.empty(),
     val attributeModifiers: List<AttributeEntry> = emptyList()
 ) {
@@ -91,12 +90,6 @@ data class Gene(
 
     private val requiredGenes: MutableSet<Holder<Gene>> = mutableSetOf()
 
-    fun isMutation(registries: HolderLookup.Provider): Boolean {
-        return GeneRegistry.getAllGeneHolders(registries)
-            .anyMatch { it.value().mutatesInto.getOrNull()?.value() === this }
-            .not()
-    }
-
     fun addRequiredGenes(genes: Collection<Holder<Gene>>) {
         requiredGenes.addAll(genes)
     }
@@ -124,7 +117,7 @@ data class Gene(
         val color = if (isActive) {
             if (isNegative) {
                 ChatFormatting.RED
-            } else if (isMutation(registries)) {
+            } else if (isMutation) {
                 ChatFormatting.DARK_PURPLE
             } else {
                 ChatFormatting.GRAY
@@ -244,6 +237,7 @@ data class Gene(
                 instance.group(
                     Codec.BOOL.optionalFieldOf("negative", false).forGetter(Gene::isNegative),
                     Codec.BOOL.optionalFieldOf("hidden", false).forGetter(Gene::isHidden),
+                    Codec.BOOL.optionalFieldOf("mutation", false).forGetter(Gene::isMutation),
                     Codec.INT.optionalFieldOf("dna_points_required", 1).forGetter(Gene::dnaPointsRequired),
                     RegistryCodecs
                         .homogeneousList(Registries.ENTITY_TYPE)
@@ -252,8 +246,6 @@ data class Gene(
                             AnyHolderSet(BuiltInRegistries.ENTITY_TYPE.asLookup())
                         )
                         .forGetter(Gene::allowedEntities),
-                    RegistryFixedCodec.create(GeneRegistry.GENE_REGISTRY_KEY).optionalFieldOf("mutates_into")
-                        .forGetter(Gene::mutatesInto),
                     PotionDetails.DIRECT_CODEC.optionalFieldOf("potion_details").forGetter(Gene::potionDetails),
                     AttributeEntry.DIRECT_CODEC.listOf().optionalFieldOf("attribute_modifiers", emptyList())
                         .forGetter(Gene::attributeModifiers)
@@ -263,9 +255,9 @@ data class Gene(
         val DIRECT_STREAM_CODEC: StreamCodec<RegistryFriendlyByteBuf, Gene> = NeoForgeStreamCodecs.composite(
             ByteBufCodecs.BOOL, Gene::isNegative,
             ByteBufCodecs.BOOL, Gene::isHidden,
+            ByteBufCodecs.BOOL, Gene::isMutation,
             ByteBufCodecs.INT, Gene::dnaPointsRequired,
             ByteBufCodecs.holderSet(Registries.ENTITY_TYPE), Gene::allowedEntities,
-            ByteBufCodecs.optional(Gene.STREAM_CODEC), Gene::mutatesInto,
             ByteBufCodecs.optional(PotionDetails.DIRECT_STREAM_CODEC), Gene::potionDetails,
             AttributeEntry.DIRECT_STREAM_CODEC.apply(ByteBufCodecs.list()), Gene::attributeModifiers,
             ::Gene