+++ CHANGELOG.md
@@ -180,9 +180,9 @@
 - They follow the following structure, all fields are optional:
     - `dna_points_required`: The amount of DNA Points required to complete a Plasmid. Defaults to 0.
     - `negative`: If the Gene is negative. Defaults to false.
+    - `mutation`: If the Gene is a mutation. Defaults to false.
     - `hidden`: If the Gene is hidden, only used by default by the Basic Gene. Defaults to false.
     - `allowed_entities`: A filter of entity types that the Gene can be injected into. Defaults to `{"type": "neoforge:any"}`. Can also accept a single entity type (`"minecraft:player"`) or a list of entity types (`["minecraft:cow","minecraft:chicken"]`)
-    - `mutates_into`
     - `potion_details`: An object with the following fields, for a potion effect that's granted at all times:
         - `effect`: The effect to apply
         - `level`: The level of the effect (optional, defaults to 1)